refactor: package reidentification
