+++ pkg/reidentification/model.go
@@ -1,0 +1,22 @@
+// Copyright (C) 2022 CGI France
+//
+// This file is part of SIGO.
+//
+// SIGO is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// SIGO is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with SIGO.  If not, see <http://www.gnu.org/licenses/>.
+
+package reidentification
+
+type Distance interface {
+	Compute(map[string]float64, map[string]float64) float64
+}
